+++ src/app/portfolio/portfolio.component.ts
@@ -227,25 +227,46 @@ export class PortfolioComponent implements OnInit {
 // }
 
 deleteProject(projectId: number): void {
+
+  Swal.fire({
+    title: "Are you sure?",
+    text: "You won't be able to revert this!",
+    icon: "warning",
+    showCancelButton: true,
+    confirmButtonColor: "#3085d6",
+    cancelButtonColor: "#d33",
+    confirmButtonText: "Yes, delete it!"
+  }).then((result) => {
+    if (result.isConfirmed) {
+
+      this.ds.deleteProject(projectId).subscribe(
+        (response) => {
+            console.log('Accomplishment deleted successfully:', response);
+            // Reload the portfolio to reflect changes
+            this.loadPortfolio();
+            Swal.fire({
+              title: "Deleted Successfully",
+              icon: "success"
+            });
+        },
+        (error) => {
+            console.error('Error deleting accomplishment:', error);
+            if (error.status === 401) {
+                console.warn('Unauthorized access - redirecting to login');
+                this.route.navigateByUrl('/login'); // Or your login route
+            }
+        }
+    );
+
+      Swal.fire({
+        title: "Deleted!",
+        text: "Your Project has been deleted.",
+        icon: "success"
+      });
+    }
+  });
   
-  this.ds.deleteProject(projectId).subscribe(
-      (response) => {
-          console.log('Accomplishment deleted successfully:', response);
-          // Reload the portfolio to reflect changes
-          this.loadPortfolio();
-          Swal.fire({
-            title: "Deleted Successfully",
-            icon: "success"
-          });
-      },
-      (error) => {
-          console.error('Error deleting accomplishment:', error);
-          if (error.status === 401) {
-              console.warn('Unauthorized access - redirecting to login');
-              this.route.navigateByUrl('/login'); // Or your login route
-          }
-      }
-  );
+  
 
 }
